[BUG] Make the Shader class non-copyable
This commit fixes a bug in the Shader class for future commits where
copies of a shader object leads to double-free or possibly orphan GPU
memory. By ensuring that the Shader is non-copyable and adding safe
checks on move assignment, we can avoid some of the most frequent memory
corruptions or leaks.

Double checking the shaderProgram or ID before deleting is another
safety mechanism to ensure we don't double delete. This is added on top
of initialising the shader, fragment and vertex identifiers to 0.