[ENH] New class: ShaderManager
Many components of the Engine and the clients rely on reusing the same
shader resources. Instead of passing shaders across the project or
storing multiple references to a single shader object on the GPU, the
ShaderManager owns the C++ Shader objects and lets users fetch or
interact with them via its APIs.