feat: new background with parallax, support textures
Add support for textures to DrawableComponent. By default, a white
texture is assigned. The background will move with the ship by a factor.
The closer to zero it is, the farther away we can simulate a background
layer to be in relation with the player.

Texture files are not currently included due to copyrights